+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `gutterPixels` options to support margins out of the box
 - Change column calculation logic to use `Math.floor` instead of `Math.round` (fixes overflowing elements at the grid's end)
+- Add new public method `removeItem` that removes an item from the grid
 
 ## Version 2.1.1